requirements.txt
bcrypt
bitsandbytes
faiss-cpu
langchain
langchain-community
langchain-experimental
langchain-openai
mimetypes
pandas
pypdf2
Pillow
python-dotenv
PyYAML
streamlit
torch
transformers
unstructured[all]
timm
